Fixed the stupid "order matters" bug
The old STG Prelude has a comment at the end of the file.  The syntax
conversion program preserves comments, but does not add newlines to
the end for prettier formatting. The resulting new Prelude then had a
comment with no newline at the end of the file.  When concatenating
input files for testing, this had the effect of commenting out the
first line of whatever followed the Prelude.

I suppose that's what I get for trying to preserve comments.