Constructor application translation done-ish
I've convinced myself that any time an StgRhsCon or a ConApp appears,
it's safe to make a CON.  GHC does make wrapper functions for newtypes
and for constructors with strict fields, and we'll attempt to honor
that.  It does not appear to be difficult to work around that if
necessary though.