Big FromGHC refactor. Lists appear to work now.
Pretty printing for GHC's STG is moved into PrettySTG.hs. Naming
logic (which depends on BuiltIn.hs) is in Naming.hs.  I may pull out the
G2A logic as well and leave FromGHC as the controller for GHC and the
main interface.

Naming is a little different now.  It turns out that we don't need
analagous datatypes for GHC's wired-ins in our base libraries. We can
always construct TyCons and DataCons based on the usage of GHC DataCons
in the STG and simply use the names they already have.  This makes
naming much simpler.  We'll see if this holds for typeclasses...

Lists appear to work, "main = [(),(),()]" evaluates as expected, but
there's no support in the Base libraries yet. That's next in line
though.