simple Eq typeclass tests seem to work!
The typeclass mapping seems to work, albeit with some hacks.  This is
only tested with Eq on Ints, Chars and Tuples thereof, but, in
principle, the rest of the classes should work as well.

This commit also includes a Tuple Eq instance "deriver" (just generates
Haskell syntax for tuples of arity `n`).  An Ord "deriver" should be
similarly simple to implement.