uniquify done, Case now has default option and scrutinee binding
On Kei's suggestion, case expressions have a binding for the scrutinee.
Following GHC's STG, there is now also a `Default` constructor for the `Clause`
type.  If/when a concrete syntax emerges, single variable bindings will likely
not be permitted. Instead, underscores will be used to indicate a default, and
the existing binding for the scrutinee can be used.